Lab 1 Tyulin Igor

+++ src/task3.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 unsigned long long sumPrime(unsigned int hbound)
 {
-    unsigned int sum = 0;
+    unsigned long long sum = 0;
     for(unsigned int i = 2; i < hbound; i++)
     {
         sum += nPrime(i) ? i : 0;
